Merge origin/plan-00066-ccy-ci-runner into F44 (post Plan 00089, 808ec31),
preserving ancestry (--no-ff). 90 branch commits land: the
_planlib.inc.bash plan-scripting library and its test suite, six new
.claude/rules/*.md pointer files, a docs-link-checker QA gate
(scripts/qa-docs.bash + helpers/docs/link_check.py), engine_assert_rootless
(Plan 00072, CCY 3.44.0 - refuses to launch when the container engine's
rootless posture is UNKNOWN rather than inferring safety), a fixed
legacy-grub cgroup check (Plan 00074), and several already-completed plans.

89 of 107 changed files auto-merged clean. 18 conflicted and were resolved
by reading both sides' intent rather than picking one wholesale - full
per-file breakdown in CLAUDE/Plan/00090-resync-ccy-ci-runner-branch-onto-f44/
PLAN.md Tasks 2.1-2.2 and the JOURNAL. ssh-handling.bash did not newly
conflict: it auto-merged clean, so Plan 00089's GH_TOKEN precedence guard
survives untouched.

An infra-reviewer pass over the 18-file resolution found and this commit
fixes: ruff.toml briefly self-contradicted (a branch-era header claiming
ruff's unenumerated 413-rule default alongside a body that enforces an
explicit 59-rule select, plus an inert clip-scan BLE001 ignore for a rule
that was never actually enabled under that select list) - corrected in
ruff.toml, scripts/qa-python.bash, and the new .claude/rules/qa-gates.md;
a harmless duplicated 4-line hunk in run.bash; and a second run.bash
version collision my own renumbering had missed, which had silently
dropped Plan 00065's server-recommended changelog entry rather than
renumbering it clear - recovered from git history and reinserted, bumping
RUN_BASH_VERSION to 1.17.0.

./scripts/qa-all.bash: exit 0, 619 files checked, including the union of
both sides' QA stages (qa-docs.bash: 56 files; qa-helper-tests.bash: 203
tests, up from 161 pre-merge).

Refs: CLAUDE/Plan/00090-resync-ccy-ci-runner-branch-onto-f44

+++ helpers/containerwatch/core.py
@@ -255,7 +255,7 @@ def make_finding(
         "argv0": argv0,
         "cmd": cmd,
         "age_s": age_s,
-        "cpu_pct": int(round(cpu_pct)),
+        "cpu_pct": round(cpu_pct),
         "rss_kb": rss_kb,
         "exec_hint": exec_hint,
     }
